MPI-related changes. Works on NERSC
Changes related to MPI.  File mpilib.f90 is no longer needed. Now all insertions are in mpins_par.f.  Also updated makefiles for NERSC/Cori.  Script doparallel.py needs some improvement for procedure of cleaning the write(*,*) lines, but it is not important for results of runs.  Small change in netcdfrw2 - works on Bob's gfortran and on NERSC.